requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
redis==6.4.0
passlib==1.7.4
bcrypt==4.0.1
aiosmtplib==3.0.1
email-validator==2.3.0
PyJWT==2.8.0
requests==2.32.3
celery==5.5.3